requirements.txt
fastapi
uvicorn
huggingface-hub
python-dotenv
python-multipart
edge-tts
